requirement.txt
Flask
Flask.Core

llama-index==0.11.13
llama-index-cli==0.3.1
llama-index-core==0.11.13.post1
llama-index-llms-ollama==0.3.2
llama-index-embeddings-ollama==0.3.1
llama-index-llms-groq

numpy==1.26.4
pandas==2.2.3
matplotlib==3.9.2
matplotlib.inline==0.1.7

python-dotenv
SQLAlchemy
